Moved exceptions to the public area for proper documentation
