+++ src/yajaw/core/__init__.py
@@ -3,4 +3,4 @@
 # SPDX-License-Identifier: MIT
 """File __init__.py responsible for enabling the import of yajaw.core package."""
 
-__all__ = ["rest", "exceptions"]
+__all__ = ["rest"]
