+++ src/yajaw/core/rest.py
@@ -7,7 +7,7 @@ from http import HTTPStatus
 import httpx
 
 from yajaw import Option, YajawConfig
-from yajaw.core import exceptions
+from yajaw import exceptions
 
 
 class PersonalAccessTokenAuth(httpx.Auth):
